refactor: Separate useFetch from useMultiframe
